+++ src/Camera.h
@@ -7,17 +7,17 @@ class Camera
 private:
     int m_hSize;
     int m_vSize;
-    float m_fov;
+    double m_fov;
     Mat4 m_transform;
-    float m_pixelSize;
-    float m_halfWidth;
-    float m_halfHeight;
+    double m_pixelSize;
+    double m_halfWidth;
+    double m_halfHeight;
 
     void CalculatePixelSize();
 
 public:
     // Constructors
-    Camera(int hSize = 160, int vSize = 120, float fieldOfView = Utils::GetPI() / 2.0f);
+    Camera(int hSize = 160, int vSize = 120, double fieldOfView = Utils::GetPI() / 2.0);
 
 
     // Destructor
@@ -27,11 +27,11 @@ public:
     // Accessors
     const int GetHSize() const;
     const int GetVSize() const;
-    const float GetFov() const;
+    const double GetFov() const;
     const Mat4 GetTransform() const;
-    const float GetPixelSize() const;
-    const float GetHalfWidth() const;
-    const float GetHalfHeight() const;
+    const double GetPixelSize() const;
+    const double GetHalfWidth() const;
+    const double GetHalfHeight() const;
 
 
     // Member functions
